# feat: add programmatic API and performance improvements